requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.1
Django==3.1.6
django-extensions==3.1.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.24
numpy==1.20.2
packaging==20.9
pandas==1.2.4
Pillow==8.1.2
pipenv==2020.11.15
pluggy==0.13.1
py==1.10.0
PyJWT==1.7.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
webdriver-manager==3.4.2
wrapt==1.12.1